ew-combat-system/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ew-combat-system"
version = "1.0.0"
description = "专业的电子战体系对抗仿真与评估平台"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "电子战仿真实验室", email = "support@ew-simulation.com"}
]
maintainers = [
    {name = "技术团队", email = "dev@ew-simulation.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Framework :: Streamlit",
]
dependencies = [
    "streamlit>=1.28.0",
    "geoviews>=1.9.1",
    "holoviews>=1.16.0",
    "bokeh>=3.3.0",
    "plotly>=5.18.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0",
    "pydantic>=2.4.0",
    "geopandas>=0.13.0",
    "shapely>=2.0.0",
    "folium>=0.14.0",
    "cartopy>=0.21.0",
    "matplotlib>=3.7.0",
    "xarray>=2023.1.0",
    "pyproj>=3.6.0",
    "colorcet>=3.0.0",
    "panel>=1.3.0",
    "param>=2.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "openpyxl>=3.1.0",
    "reportlab>=4.0.0",
    "click>=8.1.0",
    "scikit-learn>=1.3.0",
    "networkx>=3.1",
    "pydeck>=0.8.0",
    "hvplot>=0.8.0",
    "opencv-python>=4.8.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "streamlit-folium>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pylint>=3.0.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "pytest-mock>=3.11.0",
    "pytest-asyncio>=0.21.0",
    "pre-commit>=3.5.0",
    "bump2version>=1.0.0",
]
docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
    "sphinx-autodoc-typehints>=2.0.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "pytest-mock>=3.11.0",
]
notebook = [
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "ipywidgets>=8.0.0",
]

[project.scripts]
ew-sim = "ew_combat_system.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/ew-combat-system"
Documentation = "https://ew-simulation.com/docs"
Repository = "https://github.com/yourusername/ew-combat-system.git"
Issues = "https://github.com/yourusername/ew-combat-system/issues"
Changelog = "https://github.com/yourusername/ew-combat-system/blob/main/CHANGELOG.md"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=src --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
ignore_errors = true

[project.entry-points."streamlit"]
ew_simulation = "ew_combat_system.app:main"